Module aws_sdk_finspace::error

Modules

access_denied_exception

See AccessDeniedException

internal_server_exception

See InternalServerException

invalid_request_exception

See InvalidRequestException

limit_exceeded_exception

See LimitExceededException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

CreateEnvironmentError

DeleteEnvironmentError

GetEnvironmentError

InternalServerException

The request processing has failed because of an unknown error, exception or

InvalidRequestException

The request is invalid. Something is wrong with the input to the request.

LimitExceededException

A service limit or quota is exceeded.

ListEnvironmentsError

ListTagsForResourceError

ResourceNotFoundException

One or more resources can't be found.

ServiceQuotaExceededException

You have exceeded your service quota. To perform the requested action,

TagResourceError

ThrottlingException

The request was denied due to request throttling.

UntagResourceError

UpdateEnvironmentError

ValidationException

The input fails to satisfy the constraints specified by an AWS service.

Enums

CreateEnvironmentErrorKind

DeleteEnvironmentErrorKind

GetEnvironmentErrorKind

ListEnvironmentsErrorKind

ListTagsForResourceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateEnvironmentErrorKind